requirements.txt

pydantic
protobuf
pydantic-settings
python-dotenv
fastapi